requirements.txt
anyio                     3.6.2
cachetools                5.3.0
certifi                   2022.12.7
cffi                      1.15.1
charset-normalizer        3.1.0
cryptography              40.0.1
google-api-core           2.11.0
google-auth               2.16.3
google-cloud-texttospeech 2.14.1
googleapis-common-protos  1.59.0
grpcio                    1.51.3
grpcio-status             1.51.3
h11                       0.14.0
h2                        4.1.0
hpack                     4.0.0
httpcore                  0.16.3
httpx                     0.23.3
hyperframe                6.0.1
idna                      3.4
paypalrestsdk             1.13.1
pip                       23.0.1
proto-plus                1.22.2
protobuf                  4.22.1
pyasn1                    0.4.8
pyasn1-modules            0.2.8
pycparser                 2.21
pydub                     0.25.1
pyOpenSSL                 23.1.0
pyTelegramBotAPI          4.10.0
requests                  2.28.2
rfc3986                   1.5.0
rsa                       4.9
setuptools                67.4.0
six                       1.16.0
sniffio                   1.3.0
urllib3                   1.26.15
wheel                     0.38.4

